requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.12.2
black==23.3.0
cachetools==5.3.0
certifi==2022.12.7
cfgv==3.3.1
chardet==5.1.0
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
coverage==7.2.3
cssselect==1.2.0
dataclasses-json==0.5.7
distlib==0.3.6
dnspython==2.3.0
docker==6.0.1
duckduckgo-search==2.8.6
exceptiongroup==1.1.1
faiss-cpu==1.7.4
filelock==3.11.0
flake8==6.0.0
frozenlist==1.3.3
future==0.18.2
google-api-core==2.11.0
google-api-python-client==2.85.0
google-auth==2.17.3
google-auth-httplib2==0.1.0
google-search-results==2.4.2
googleapis-common-protos==1.59.0
greenlet==2.0.2
gTTS==2.3.1
h11==0.14.0
httplib2==0.22.0
identify==2.5.22
idna==3.4
isort==5.12.0
langchain==0.0.162
loguru==0.7.0
lxml==4.9.2
macholib==1.15.2
marshmallow==3.19.0
marshmallow-enum==1.5.1
mccabe==0.7.0
multidict==6.0.4
mypy-extensions==1.0.0
nodeenv==1.7.0
numexpr==2.8.4
numpy==1.24.2
openai==0.27.4
openapi-schema-pydantic==1.2.4
orjson==3.8.10
outcome==1.2.0
packaging==23.1
pandas==2.0.1
pathspec==0.11.1
Pillow==9.5.0
pinecone-client==2.2.1
pip==23.1
platformdirs==3.2.0
playsound==1.2.2
pre-commit==3.2.2
protobuf==4.22.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pydantic==1.10.7
pyflakes==3.0.1
pyparsing==3.0.9
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
PyYAML==6.0
readability-lxml==0.8.1
redis==4.5.4
regex==2023.3.23
requests==2.28.2
rsa==4.9
selenium==4.8.3
setuptools==58.0.4
six==1.15.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
sourcery==1.2.0
SQLAlchemy==2.0.12
tenacity==8.2.2
tiktoken==0.3.3
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==1.26.15
virtualenv==20.21.0
webdriver-manager==3.8.6
websocket-client==1.5.1
wheel==0.37.0
wsproto==1.2.0
yapf==0.33.0
yarl==1.8.2
